@@ -295,8 +295,8 @@
     "_mocha": "./bin/_mocha"
   },
   "engines": {
-    "node": ">= 0.10.x",
-    "npm": ">= 1.4.x"
+    "node": ">= 4.0.0",
+    "npm": ">= 2.14.2"
   },
   "scripts": {
     "lint": "eslint . bin/*",
@@ -307,11 +307,11 @@
   "dependencies": {
     "browser-stdout": "1.3.0",
     "commander": "2.9.0",
-    "debug": "2.6.8",
+    "debug": "2.6.9",
     "diff": "3.2.0",
     "escape-string-regexp": "1.0.5",
     "glob": "7.1.1",
-    "growl": "1.9.2",
+    "growl": "1.10.2",
     "he": "1.1.1",
     "json3": "3.3.2",
     "lodash.create": "3.1.1",
@@ -325,11 +325,13 @@
     "coffee-script": "^1.10.0",
     "coveralls": "^2.11.15",
     "cross-spawn": "^5.1.0",
-    "eslint": "^3.11.1",
-    "eslint-config-semistandard": "^7.0.0",
-    "eslint-config-standard": "^6.2.1",
+    "eslint": "^4.7.2",
+    "eslint-config-semistandard": "^11.0.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-node": "^5.1.1",
     "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-standard": "2.0.1",
+    "eslint-plugin-standard": "^3.0.1",
     "expect.js": "^0.3.1",
     "karma": "1.3.0",
     "karma-browserify": "^5.0.5",
@@ -337,13 +339,13 @@
     "karma-expect": "^1.1.2",
     "karma-mocha": "^1.3.0",
     "karma-phantomjs-launcher": "0.2.3",
-    "karma-spec-reporter": "0.0.26",
+    "karma-spec-reporter": "0.0.31",
     "nyc": "^11.2.1",
     "os-name": "^2.0.1",
     "phantomjs": "1.9.8",
     "readable-stream": "2.2.11",
     "rimraf": "^2.5.2",
-    "should": "^9.0.2",
+    "should": "^13.1.0",
     "through2": "^2.0.1",
     "watchify": "^3.7.0"
   },
@@ -359,8 +361,7 @@
     "bower.json"
   ],
   "browser": {
-    "debug": "./lib/browser/debug.js",
-    "events": "./lib/browser/events.js",
+    "growl": "./lib/browser/growl.js",
     "tty": "./lib/browser/tty.js",
     "./index.js": "./browser-entry.js",
     "fs": false,
