@@ -120,8 +120,8 @@ exports.window = {
 
 if (isatty) {
   exports.window.width = process.stdout.getWindowSize
-      ? process.stdout.getWindowSize(1)[0]
-      : tty.getWindowSize()[1];
+    ? process.stdout.getWindowSize(1)[0]
+    : tty.getWindowSize()[1];
 }
 
 /**
