@@ -229,7 +229,7 @@ Runner.prototype.fail = function (test, err) {
   ++this.failures;
   test.state = 'failed';
 
-  if (!(err instanceof Error || err && typeof err.message === 'string')) {
+  if (!((err instanceof Error || err) && typeof err.message === 'string')) {
     err = new Error('the ' + type(err) + ' ' + stringify(err) + ' was thrown, throw an Error :)');
   }
 
@@ -722,7 +722,7 @@ Runner.prototype.uncaught = function (err) {
     return;
   }
 
- // recover from hooks
+  // recover from hooks
   if (runnable.type === 'hook') {
     var errSuite = this.suite;
     // if hook failure is in afterEach block
