@@ -540,8 +540,8 @@ function jsonStringify (object, spaces, depth) {
     --length;
     str += '\n ' + repeat(' ', space) +
       (isArray(object) ? '' : '"' + i + '": ') + // key
-      _stringify(object[i]) +                    // value
-      (length ? ',' : '');                       // comma
+      _stringify(object[i]) + // value
+      (length ? ',' : ''); // comma
   }
 
   return str +
